almetrics
Altmetrics is the study and use of non-traditional scholarly impact measures that are based on activity in web-based environments. As scholarship increasingly moves online, these metrics track associated interactions and activity to generate fine-grained data, allowing researchers and policy makers to create a higher resolution picture of the reach and impact of academic research. In thei workshop ImpactStory, one of the leading open source Altmetrics services will be demonstrated.